requirements.txt
paramiko==3.5.0
python-dotenv==1.0.1
Requests==2.32.3
python-whois==0.9.5